modules/db/pyproject.toml
[project]
name = "angular"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.121.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "regex>=2025.11.3",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
